+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Fix redundant public accessibility analysis false-positive for actors. 
 - Fix redundant public accessibility analysis false-positive for property wrappers.
 - Fix redundant public accessibility analysis false-positive for declarations referenced from a public `@inlinable` function.
+- Fix redundant public accessibility analysis false-positive for function parameter default values.
 
 ## 2.17.1 (2023-12-04)